Update description in description.html
The descriptions of the main features and key attributes have been revamped in the description.html file. This update makes the information about the IDE Plugin for Nginx Configuration Files more comprehensible and attractive to potential users. Changes also involved enhancing the details for each feature and providing a clear list of key features.